Add block to removeApp.

+++ app/Http/Controllers/GigController.php
@@ -514,6 +514,12 @@ class GigController extends Controller
 
     public function removeApp(Job $job)
     {
+        $bookedUserID = $job->users()->select(['users.*'])->wherePivot('status', 'Booked')->first()->id ?? '';
+
+        if ($bookedUserID == Auth::id()) {
+            return redirect()->back()->with('warning', 'You\'ve been booked for this gig and only the host can remove your application');
+        }
+
         $job->users()->detach(Auth::id());
 
         return redirect()->back()->with('success', 'You\'ve successfully removed your application.');
